refactor(ts-oss): dedupe toCamelCasePreservingIds into shared utils
redis.ts and valkey.ts each defined an identical toCamelCasePreservingIds
helper. Move it to utils/casing.ts, alongside the other single-purpose
oss utils that vector stores already import (e.g. utils/load_peer),
and import it from both stores.
