+++ mem0-ts/src/oss/src/vector_stores/valkey.ts
@@ -2,6 +2,7 @@ import { VectorStore } from "./base";
 import { SearchFilters, VectorStoreResult } from "../types";
 import { ValkeyConfig } from "../types/valkey";
 import { loadPeer } from "../utils/load_peer";
+import { toCamelCasePreservingIds } from "../utils/casing";
 
 interface ValkeyClient {
   call: (...args: (string | number | Buffer)[]) => Promise<unknown>;
@@ -43,28 +44,6 @@ function toSnakeCase(obj: Record<string, any>): Record<string, any> {
       value,
     ]),
   );
-}
-
-function toCamelCase(obj: Record<string, any>): Record<string, any> {
-  if (typeof obj !== "object" || obj === null) return obj;
-  return Object.fromEntries(
-    Object.entries(obj).map(([key, value]) => [
-      key.replace(/_([a-z])/g, (_, letter) => letter.toUpperCase()),
-      value,
-    ]),
-  );
-}
-
-function toCamelCasePreservingIds(
-  payload: Record<string, any>,
-): Record<string, any> {
-  const { agent_id, run_id, user_id, ...rest } = payload;
-  return {
-    ...toCamelCase(rest),
-    ...(agent_id !== undefined && { agent_id }),
-    ...(run_id !== undefined && { run_id }),
-    ...(user_id !== undefined && { user_id }),
-  };
 }
 
 interface ValkeySearchDoc {
